+++ src/or/replaycache.h
@@ -46,7 +46,7 @@ replaycache_t * replaycache_new(time_t horizon, time_t interval);
  */
 
 STATIC int replaycache_add_and_test_internal(
-    time_t present, replaycache_t *r, const void *data, int len,
+    time_t present, replaycache_t *r, const void *data, size_t len,
     time_t *elapsed);
 STATIC void replaycache_scrub_if_needed_internal(
     time_t present, replaycache_t *r);
@@ -57,9 +57,9 @@ STATIC void replaycache_scrub_if_needed_internal(
  * replaycache_t methods
  */
 
-int replaycache_add_and_test(replaycache_t *r, const void *data, int len);
+int replaycache_add_and_test(replaycache_t *r, const void *data, size_t len);
 int replaycache_add_test_and_elapsed(
-    replaycache_t *r, const void *data, int len, time_t *elapsed);
+    replaycache_t *r, const void *data, size_t len, time_t *elapsed);
 void replaycache_scrub_if_needed(replaycache_t *r);
 
 #endif
